backend/requirements.txt
# backend/requirements.txt

fastapi
uvicorn
psycopg2-binary
pydantic
prometheus_client